+++ src/pas/plugins/headers/plugins.py
@@ -124,8 +124,9 @@ class HeaderPlugin(BasePlugin):
                 b'ERROR: denying any unauthorized access.\n')
             return True
         if self.redirect_url:
-            logger.debug('Redirecting to %s', self.redirect_url)
-            response.redirect(self.redirect_url, lock=1)
+            url = '{}?came_from={}'.format(self.redirect_url, request.URL)
+            logger.warning('Redirecting to %s', url)
+            response.redirect(url, lock=1)
             return True
         # We have no redirect_url, so we do not know how to challenge.
         # Let Plone handle this in the default way, probably showing